magit-clone: Add --filter to support partial clones
Git gained partial clone support via its --filter argument in v2.17
(undocumented until v2.27.0).  Add the option to the clone transient,
along with a custom reader that offers the two mostly likely values of
interest (blobless and treeless clones).  See git-rev-list's manpage
for a complete description of the filter specification.

Note that users should probably expect to hit into rough edges when
working with a partial clone.  For example, without a connection,
checking out a commit for which blobs need to be downloaded will
"fail" in the sense that it the checkout command will emit errors and
`git status' from the command line will show deleted files for files
that couldn't be downloaded.  While that's not surprising, the
checkout command has an exit status of 0 [*], so Magit users won't see
an indication of the failure.  And worse, the deleted files won't show
up in the status buffer because the underlying git-diff commands fail
(for the same reason as the checkout: the required objects can't be
downloaded).

[*] I haven't yet reported this upstream, but it seems like a bug to
    me.