sparse-checkout: Add insert-header function
It's useful to have some indication that you're in a sparse checkout.
Add a function that inserts a header similar in style to the "Filter!"
header inserted by magit-insert-diff-filter-header.

It may end up being useful to add this to magit-status-headers-hook by
default, in which case we'll probably need to rework where things are
defined.  But for now, just add the function and let interested users
add it to magit-status-headers-hook.