magit-log: Fix order in which suffixes are displayed
The "local branches" and "all branches" variants are obviously closely
related and should not be separated by much less related variants.

The "current" and "HEAD" variants are even more related and should not
be separated by the very much not related "other" variant.

Also change the order of the command definitions and in the manual.