Rename magit-utils.el to magit-base.el
It hasn't been true for quite some time that this library only (or
even just mostly) defined things that have "very little to do with
Git or Magit", and it is time to account for that.

"magit-base.el" serves a role similar to that of "magit-core.el",
except that it comes even earlier in the dependency tree.  For that
reason other libraries that depend on one or more of the libraries
that `magit-base' depends on should just require that feature.