magit-log-show-signatures-limit: New option
Checking the signature of more than a few hundred commits is just
too slow.  It would be nice if "--show-signature" took a numeric
value, then we could at least verify the signatures of the last N
commits.  Because it does not, we cannot verify any commits, when
showing a large number of commits.